backend/requirements.txt
fastapi
uvicorn
python-binance
pandas<3.0.0
pandas-ta
python-dotenv
pydantic
websockets
python-multipart
httpx
